Consolidate Redux raid and dungeon actions into instance actions
